Factor out common write operations for @propertysheets POST.
