+++ opengever/propertysheets/api/post.py
@@ -1,18 +1,12 @@
-from opengever.api.validation import get_validation_errors
-from opengever.api.validation import scrub_json_payload
-from opengever.propertysheets.api.base import PropertySheetLocator
-from opengever.propertysheets.definition import PropertySheetSchemaDefinition
+from opengever.propertysheets.api.base import PropertySheetWriter
 from opengever.propertysheets.exceptions import InvalidSchemaAssignment
-from opengever.propertysheets.metaschema import IFieldDefinition
-from plone import api
 from plone.protect.interfaces import IDisableCSRFProtection
 from plone.restapi.deserializer import json_body
 from zExceptions import BadRequest
-from zExceptions import Unauthorized
 from zope.interface import alsoProvides
 
 
-class PropertySheetsPost(PropertySheetLocator):
+class PropertySheetsPost(PropertySheetWriter):
     """
     Add new property sheets or replace existing ones.
 
@@ -63,52 +57,3 @@ class PropertySheetsPost(PropertySheetLocator):
 
         self.request.response.setStatus(201)
         return self.serialize(schema_definition)
-
-    def validate_fields(self, fields):
-        errors = []
-
-        for field_data in fields:
-            # Cast JSON strings to their appropriate Python types (unicode or
-            # bytestring), depending on the schema field (ASCII or Text[Line])
-            scrub_json_payload(field_data, IFieldDefinition)
-
-            # Allowing unknown fields is necessary in order to allow managers
-            # to set dynamic defaults like `default_factory`, which currently
-            # aren't exposed in the meta schema.
-            field_errors = get_validation_errors(
-                field_data, IFieldDefinition, allow_unknown_fields=True)
-
-            if field_errors:
-                errors.extend(field_errors)
-
-            # Require Manager role for any kind of dynamic defaults
-            dynamic_defaults = PropertySheetSchemaDefinition.DYNAMIC_DEFAULT_PROPERTIES
-            if any(p in field_data for p in dynamic_defaults):
-                if not api.user.has_permission('cmf.ManagePortal'):
-                    raise Unauthorized(
-                        'Setting any dynamic defaults requires Manager role')
-        return errors
-
-    def create_property_sheet(self, sheet_name, assignments, fields):
-        schema_definition = PropertySheetSchemaDefinition.create(
-            sheet_name, assignments=assignments
-        )
-        for field_data in fields:
-            name = field_data["name"]
-            field_type = field_data["field_type"]
-            title = field_data.get("title", name.decode('ascii'))
-            description = field_data.get("description", u"")
-            required = field_data.get("required", False)
-            values = field_data.get("values")
-            default = field_data.get("default")
-            default_factory = field_data.get("default_factory")
-            default_expression = field_data.get("default_expression")
-            default_from_member = field_data.get("default_from_member")
-            schema_definition.add_field(
-                field_type, name, title, description, required, values,
-                default, default_factory, default_expression,
-                default_from_member
-            )
-
-        self.storage.save(schema_definition)
-        return self.storage.get(sheet_name)
